backend/requirements.txt
fastapi
uvicorn
aiofiles
python-multipart
PyPDF2
openai>=1.0.0
azure-search-documents
azure-storage-blob
python-dotenv
aiohttp>=3.8.0
azure-search-documents
tiktoken